also run handleRes after updateMulti, add possibility of returned items in destroy

+++ src/api.js
@@ -59,10 +59,6 @@ const handleRes = (res, kind, params = {}) => {
       createMultiLocal(kind, items)
       return {ok: true, items}
     }
-
-    
-
-    
     
   }
   return res;
@@ -118,17 +114,7 @@ export async function index(kind, {ignoreStamp, inverted, localKind, refresh, lo
 
 export async function updateMulti(kind, params, {localKind, extraParams, saveLocal = true} = {}) {
   const res = await backend.post(`${kind}/multi_update`, {items: params, ...extraParams})
-  if(res.ok && res.data && res.data.items) {
-    const {items} = res.data
-    saveLocal && createMultiLocal(kind, items)
-    return {ok: true, items}
-  }
-  else if(res.ok) {
-    return {ok: true}
-  }
-   else {
-    return {ok: false}
-  }
+  return handleRes(res, kind)
 }
 
 export async function show(kind, id, {extraParams = {}, localKind} = {} ) {
@@ -140,9 +126,6 @@ export async function show(kind, id, {extraParams = {}, localKind} = {} ) {
     return {ok: true, item}
   }
 }
-
-
-
 
 export const create = add;
 export const update = add;
@@ -158,7 +141,7 @@ export async function destroyMutation(kind, id) {
 export async function destroy(kind, id, extraParams = {}) {
   const url = config.deleteUrls[kind] ? funcOrValue(config.deleteUrls[kind], id) : `${kind}/${id}`
   const res = await backend.delete(url)
-  if(res.ok && !res?.data?.destroyed_ids) {
+  if(res.ok && !res?.data?.destroyed_ids && !res.data.items) {
     destroyLocal(kind, id, extraParams)
     return {ok: true}
   } else {
